Update patch set 1
Patch Set 1:

> Just symlink the files....

Firmware is under half a meg, it's a lot cleaner to just do it this way. It's not like modem firmware where we're symlinking from device, this is shipped firmware that was being symlinked with a script. If it's being shipped, it's going to be handled in vendor/

Patch-set: 1